python_backend/requirements.txt
flask-cors
click==8.0.1
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
matplotlib
pandas
nvidia-pyindex
tritonclient[all]
astroid==2.5.6
async-generator==1.10
attrdict==2.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
decorator==5.0.7
defusedxml==0.7.1
entrypoints==0.3
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
grpcio==1.37.0
importlib-metadata
ipykernel==5.5.3
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.17.0
Jinja2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.6.0
MarkupSafe
matplotlib==3.4.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
opencv-python==4.5.3.56
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==2.0.10
protobuf==3.15.8
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.8.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-rapidjson==1.0
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp
zope.event==4.5.0
zope.interface==5.4.0
requests
sklearn
scikit-image